refactor(web): extract sidebar chrome components from DesignSession
Part of issue #642 seam 5b-2 (presentational components with explicit
props). Pulls the render-only sidebar sub-blocks out of DesignSession's
`controls` closure verbatim: SessionGearMenu (tools dropdown + theme
toggle), CatalogPanel (antenna/variant picker + load-error surfaces),
DesignFreqRow, KnobOptMenu, SolverSlotTabs, and GroundPanel. Each is a
module-level component with explicit props; JSX bodies are unchanged.

gearMenuOpen stays lifted in DesignSession (downloadNec/loadMeasured
close it as a side effect), so it crosses as a prop rather than
becoming SessionGearMenu's own state.
